move test to test_forward.py
